+++ src/jest.js
@@ -121,18 +121,6 @@ function readConfig(argv, packageRoot) {
       config.useStderr = true;
     }
 
-    if (argv.testRunner) {
-      try {
-        config.testRunner = require.resolve(
-          argv.testRunner.replace(/<rootDir>/g, config.rootDir)
-        );
-      } catch (e) {
-        throw new Error(
-          `jest: testRunner path "${argv.testRunner}" is not a valid path.`
-        );
-      }
-    }
-
     if (argv.logHeapUsage) {
       config.logHeapUsage = argv.logHeapUsage;
     }
@@ -149,7 +137,7 @@ function readRawConfig(argv, packageRoot) {
   }
 
   if (typeof argv.config === 'object') {
-    return Promise.resolve(utils.normalizeConfig(argv.config));
+    return Promise.resolve(utils.normalizeConfig(argv.config, argv));
   }
 
   const pkgJsonPath = path.join(packageRoot, 'package.json');
@@ -162,7 +150,7 @@ function readRawConfig(argv, packageRoot) {
     } else {
       pkgJson.jest.rootDir = path.resolve(packageRoot, pkgJson.jest.rootDir);
     }
-    const config = utils.normalizeConfig(pkgJson.jest);
+    const config = utils.normalizeConfig(pkgJson.jest, argv);
     config.name = pkgJson.name;
     return Promise.resolve(config);
   }
@@ -171,9 +159,8 @@ function readRawConfig(argv, packageRoot) {
   return Promise.resolve(utils.normalizeConfig({
     name: packageRoot.replace(/[/\\]/g, '_'),
     rootDir: packageRoot,
-    testPathDirs: [packageRoot],
     testPathIgnorePatterns: ['/node_modules/.+'],
-  }));
+  }, argv));
 }
 
 function findOnlyChangedTestPaths(testRunner, config) {